changed back to Int64 since grammar parsing cant handle this...

+++ src/data/ARC_AGI1/arc_primitives.jl
@@ -29,7 +29,7 @@
     - Container: Grid | Object | Objects | Indices — Generic containers
 =#
 
-Integer = Int8
+Integer = Int64
 Grid = Matrix{Integer}
 GridContainer = Vector{Grid}
 IntContainer = Vector{Integer}
